refactor: seperate thread stack component
